requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile ./requirements.in
#
python-dotenv==1.0.1
    # via -r ./requirements.in
